Creating routers for color and user enpoints
