+++ models/database.py
@@ -28,3 +28,11 @@ except Exception as e:
     raise e
 
 Base = declarative_base()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
